Removing CafeteriaMenuManager
