+++ TUM Campus App/LocationProvider.swift
@@ -1,0 +1,15 @@
+//
+//  LocationProvider.swift
+//  Campus
+//
+//  Created by Mathias Quintero on 10/19/17.
+//  Copyright © 2017 LS1 TUM. All rights reserved.
+//
+
+import Foundation
+
+class LocationProvider {
+    
+    
+    
+}
