+++ TUM Campus App/TumOnlineLoginRequestManager.swift
@@ -10,6 +10,8 @@ import Sweeft
 import Alamofire
 import SWXMLHash
 
+// TODO: Refactor this
+
 class TumOnlineLoginRequestManager {
     
     init(delegate: AccessTokenReceiver?) {
